% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_rambler_loaded}
\alias{check_rambler_loaded}
\title{Check that Rambler package has loaded successfully}
\usage{
check_rambler_loaded()
}
\description{
Simple function to check that Rambler package has loaded 
  successfully. Prints "Rambler loaded successfully!" if so.
}
