% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Test function to run an example MCMC}
\usage{
run_mcmc(
  df_data,
  haplo_freqs,
  theta = 1,
  decay_rate_meanlog = 0,
  decay_rate_sdlog = 1,
  sens_shape1 = 10,
  sens_shape2 = 1,
  mu_mean = 0,
  mu_sd = 10,
  sigma_shape = 1,
  sigma_scale = 1,
  burnin = 100,
  samples = 1000,
  beta = 1,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{df_data}{data.frame of haplotypes in each individual at each time
point, in long format.}

\item{haplo_freqs}{vector giving relative frequencies of each haplotype. This
will be automatically  normalised to be relative to the maximum value.}

\item{theta}{scaling parameter on relative haplotype frequencies. The mean
number of haplotypes introduced in a single bite is
\code{theta*sum(haplo_freqs)}.}

\item{decay_rate_meanlog, decay_rate_sdlog}{mean and standard deviation (on
the log scale) of the prior distribution on the decay rate.}

\item{sens_shape1, sens_shape2}{shape parameters of beta prior on sensitivity.}

\item{mu_mean, mu_sd}{mean and standard deviation of the normal hyper-prior on
mu, which is the log-mean of the log-normal prior on lambda.}

\item{sigma_shape, sigma_scale}{shape and scale parameters of the inverse
gamma hyper-prior on sigma. The mean of this distribution is given by:
\deqn{\beta / (\alpha - 1)} for \eqn{\alpha > 1}, and the variance is given
by \deqn{\beta^2 / ((\alpha - 1)^2(\alpha - 2))} where \eqn{\alpha} is the
shape and \eqn{\beta} is the scale.}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{beta}{vector of thermodynamic powers. Final value in the vector should
always be 1.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100\%. This avoids large amounts of
output being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
This should be replaced by a more carefully thought out
  structure, probably involving defining a parameters data.frame outside of
  this function that can be loaded in.
}
