% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_cohort}
\alias{sim_cohort}
\title{Simulate a complete cohort}
\usage{
sim_cohort(n, samp_time, haplo_freqs, lambda, decay_rate, sens)
}
\arguments{
\item{n}{number of individuals in cohort.}

\item{samp_time}{vector of times at which samples were sequenced.}

\item{haplo_freqs}{vector giving the probability that each haplotype is
transmitted in a given infectious bite (not quite the same thing as
haplotype frequencies in the population).}

\item{lambda}{vector of FOI in each individual.}

\item{decay_rate}{rate at which each haplotype clears.}

\item{sens}{sensitivity of sequencing (assumed the same for all haplotypes).}
}
\description{
Simulate a cohort of individuals from the generative model. Runs
  \code{sim_ind()} multiple times, and reformats observed output into
  data.frame. Also keeps hold of the true states of all individuals for
  reference.
}
