% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Test function to run an example MCMC}
\usage{
run_mcmc(
  df_data,
  haplo_freqs,
  lambda,
  sens = 0.9,
  burnin = 100,
  samples = 1000,
  beta = 1,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{df_data}{data.frame of haplotypes in each individual at each time
point, in long format.}

\item{haplo_freqs}{vector giving the probability that each haplotype is
transmitted in a given infectious bite (not quite the same thing as
haplotype frequencies in the population).}

\item{lambda}{vector of FOI in each individual.}

\item{sens}{sensitivity of sequencing (assumed the same for all haplotypes).}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{beta}{vector of thermodynamic powers. Final value in the vector should
always be 1.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100\%. This avoids large amounts of
output being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
This should be replaced by a more carefully thought out
  structure, probably involving defining a parameters data.frame outside of
  this function that can be loaded in.
}
