% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_cohort}
\alias{sim_cohort}
\title{Simulate a complete cohort}
\usage{
sim_cohort(n, samp_time, haplo_freqs, lambda, decay_rate, sens)
}
\description{
Simulate a cohort of individuals from the generative model. Runs
  \code{sim_ind()} multiple times, and reformats observed output into
  data.frame. Also keeps hold of the true states of all individuals for
  reference.
}
