% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_ind}
\alias{sim_ind}
\title{Simulate a single individual}
\usage{
sim_ind(samp_time, haplo_freqs, lambda, decay_rate, sens)
}
\description{
Simulate a single individual from the generative model. Returns outpus as
}
